requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.20.54
botocore==1.23.54
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==3.2.6
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jwt==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-http-client==3.3.6
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.1
sendgrid==6.9.6
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==2.0.3
toml==0.10.2
twilio==7.5.1
uritemplate==4.1.1
urllib3==1.26.7
